created utility function for the time and remaining time
